code/pubsub/bq_create_views_and_cleanup/requirements.txt
google-cloud-storage
google-cloud-bigquery

